Moved ExtendedYesNoField to wagtailpgaes app. Migration included.
