+++ network-api/networkapi/wagtailpages/pagemodels/serializers.py
@@ -2,7 +2,7 @@ import datetime
 from rest_framework import serializers
 from wagtail_airtable.serializers import AirtableSerializer
 
-from networkapi.buyersguide.fields import ExtendedYesNoField
+from networkapi.wagtailpages.fields import ExtendedYesNoField
 from networkapi.wagtailpages.pagemodels.products import TRACK_RECORD_CHOICES
